fix: 효율적인 api 가독성을 위해 Controller 분리
