+++ src/main/java/com/tave/tavewebsite/domain/member/controller/ManagerController.java
@@ -20,30 +20,6 @@ public class ManagerController {
 
     private final MemberService memberService;
 
-    @PostMapping("/auth/signup")
-    public SuccessResponse<MailResponseDto> registerManager(@RequestBody @Valid RegisterManagerRequestDto requestDto) {
-        MailResponseDto response = memberService.saveMember(requestDto);
-        return new SuccessResponse<>(response);
-    }
-
-    @PostMapping("/auth/signin")
-    public SuccessResponse<SignInResponseDto> signIn(@RequestBody SignUpRequestDto requestDto) {
-        SignInResponseDto signInResponseDto = memberService.signIn(requestDto);
-        return new SuccessResponse<>(signInResponseDto);
-    }
-
-    @PostMapping("/auth/refresh")
-    public SuccessResponse<JwtToken> refreshToken(@RequestBody RefreshTokenRequestDto requestDto) {
-        JwtToken jwtToken = memberService.refreshToken(requestDto);
-        return new SuccessResponse<>(jwtToken);
-    }
-
-    @GetMapping("/auth/signout")
-    public SuccessResponse signOut(@RequestHeader("Authorization") String token) {
-        memberService.singOut(token);
-        return SuccessResponse.ok();
-    }
-
     @GetMapping("/admin/unauthorized")
     public SuccessResponse<List<UnauthorizedManagerResponseDto>> getUnauthorizedManager() {
         List<UnauthorizedManagerResponseDto> response = memberService.getUnauthorizedManager();
@@ -55,12 +31,6 @@ public class ManagerController {
         memberService.validateNickname(nickName);
         CheckNickNameResponseDto response = new CheckNickNameResponseDto(nickName);
         return new SuccessResponse<>(response, nickName + " 사용가능합니다.");
-    }
-
-    @DeleteMapping("/auth/delete/{memberId}")
-    public SuccessResponse deleteMember(@PathVariable("memberId") Long memberId) {
-        memberService.deleteMember(memberId);
-        return SuccessResponse.ok();
     }
 
     // ci/cd 이후 배포 성공 테스트용 엔드포인트
